+++ docs/CHANGE_LOG.md
@@ -18,7 +18,7 @@ Change the contents of this file to this:
 - Fixed issue with GitHub alerter raising new release alerts in reverse order for multiple releases.
 - The data store components are now compatible with the base Chainlink integration features.
 - Added Chainlink Node Alerter Manager and tests.
-- Web-Installer Bug Fixes and removing outdated alerts.
+- Web-Installer bug fixes and removing outdated alerts.
 
 ## 0.1.2